+++ src/openhuman/agent/agent/types.rs
@@ -1,0 +1,83 @@
+//! `Agent` and `AgentBuilder` struct definitions.
+//!
+//! The data shapes live here, separate from their behaviour, so the
+//! rest of the sub-module (`builder.rs`, `turn.rs`, `runtime.rs`) can
+//! focus on logic. Fields are `pub(super)` so sibling files that
+//! `impl Agent`/`impl AgentBuilder` can see them without the whole
+//! crate gaining field access.
+
+use crate::openhuman::agent::context_pipeline::ContextPipeline;
+use crate::openhuman::agent::dispatcher::ToolDispatcher;
+use crate::openhuman::agent::hooks::PostTurnHook;
+use crate::openhuman::agent::memory_loader::MemoryLoader;
+use crate::openhuman::agent::prompt::SystemPromptBuilder;
+use crate::openhuman::memory::Memory;
+use crate::openhuman::providers::{ConversationMessage, Provider};
+use crate::openhuman::tools::{Tool, ToolSpec};
+use std::sync::Arc;
+
+/// An autonomous or semi-autonomous AI agent.
+///
+/// The `Agent` is the central component that manages conversation state,
+/// executes tools based on model requests, and interacts with the memory
+/// system to maintain context across turns.
+pub struct Agent {
+    pub(super) provider: Arc<dyn Provider>,
+    pub(super) tools: Arc<Vec<Box<dyn Tool>>>,
+    pub(super) tool_specs: Arc<Vec<ToolSpec>>,
+    pub(super) memory: Arc<dyn Memory>,
+    pub(super) prompt_builder: SystemPromptBuilder,
+    pub(super) tool_dispatcher: Box<dyn ToolDispatcher>,
+    pub(super) memory_loader: Box<dyn MemoryLoader>,
+    pub(super) config: crate::openhuman::config::AgentConfig,
+    pub(super) model_name: String,
+    pub(super) temperature: f64,
+    pub(super) workspace_dir: std::path::PathBuf,
+    pub(super) identity_config: crate::openhuman::config::IdentityConfig,
+    pub(super) skills: Vec<crate::openhuman::skills::Skill>,
+    pub(super) auto_save: bool,
+    pub(super) history: Vec<ConversationMessage>,
+    pub(super) classification_config: crate::openhuman::config::QueryClassificationConfig,
+    pub(super) available_hints: Vec<String>,
+    pub(super) post_turn_hooks: Vec<Arc<dyn PostTurnHook>>,
+    pub(super) learning_enabled: bool,
+    pub(super) event_session_id: String,
+    pub(super) event_channel: String,
+    /// Layered context reduction pipeline (tool-result budget →
+    /// microcompact → autocompact signal → session-memory extraction
+    /// trigger). Owned by the agent so its state (token counters,
+    /// session-memory extraction deltas, compaction circuit breaker)
+    /// persists across turns. See
+    /// [`crate::openhuman::agent::context_pipeline`] for the stage
+    /// ordering and cache-safety contract.
+    pub(super) context_pipeline: ContextPipeline,
+}
+
+/// A builder for creating `Agent` instances with custom configuration.
+pub struct AgentBuilder {
+    pub(super) provider: Option<Arc<dyn Provider>>,
+    pub(super) tools: Option<Vec<Box<dyn Tool>>>,
+    pub(super) memory: Option<Arc<dyn Memory>>,
+    pub(super) prompt_builder: Option<SystemPromptBuilder>,
+    pub(super) tool_dispatcher: Option<Box<dyn ToolDispatcher>>,
+    pub(super) memory_loader: Option<Box<dyn MemoryLoader>>,
+    pub(super) config: Option<crate::openhuman::config::AgentConfig>,
+    pub(super) model_name: Option<String>,
+    pub(super) temperature: Option<f64>,
+    pub(super) workspace_dir: Option<std::path::PathBuf>,
+    pub(super) identity_config: Option<crate::openhuman::config::IdentityConfig>,
+    pub(super) skills: Option<Vec<crate::openhuman::skills::Skill>>,
+    pub(super) auto_save: Option<bool>,
+    pub(super) classification_config: Option<crate::openhuman::config::QueryClassificationConfig>,
+    pub(super) available_hints: Option<Vec<String>>,
+    pub(super) post_turn_hooks: Vec<Arc<dyn PostTurnHook>>,
+    pub(super) learning_enabled: bool,
+    pub(super) event_session_id: Option<String>,
+    pub(super) event_channel: Option<String>,
+}
+
+impl Default for AgentBuilder {
+    fn default() -> Self {
+        Self::new()
+    }
+}
